+++ resources/views/staff/overtimex.blade.php
@@ -9,10 +9,9 @@
     <div class="panel-body">
     
         <div class="text-center" style="margin-bottom: 15px">
-            <form action="{{route('ot.form')}}" method="POST" style="display:inline">
-                @csrf
-                <button type="submit" class="btn btn-primary">CREATE NEW CLAIM</button>
-            </form>
+            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#newOT">
+                CREATE NEW CLAIM
+            </button>
         </div>
         <div class="table-responsive">
             <table id="tOTList" class="table table-bordered">
@@ -36,7 +35,7 @@
                         <td>{{ $singleuser->status }}</td>
                         <td>
                             @if($singleuser->status=="Draft")
-                                <form action="{{route('ot.update')}}" method="POST" style="display:inline">
+                                <form action="{{route('ot.edit')}}" method="POST" style="display:inline">
                                     @csrf
                                     <input type="text" class="hidden" id="inputid" name="inputid" value="{{$singleuser->id}}" required>
                                     <button type="submit" class="btn btn-primary"><i class="fas fa-pencil-alt"></i></button>
@@ -45,7 +44,7 @@
                                     <i class="fas fa-trash-alt"></i>
                                 </button>
                             @else
-                            <form action="{{route('ot.update')}}" method="POST" style="display:inline">
+                            <form action="{{route('ot.edit')}}" method="POST" style="display:inline">
                                     @csrf
                                     <input type="text" class="hidden" id="inputid" name="inputid" value="{{$singleuser->id}}" required>
                                     <button type="submit" class="btn btn-primary"><i class="fas fa-eye"></i></button>
@@ -60,6 +59,32 @@
     </div>
 </div>
 
+<div id="newOT" class="modal fade" role="dialog">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <button type="button" class="close" data-dismiss="modal">&times;</button>
+                <h4 class="modal-title">Create New OT Claim</h4>
+            </div>
+            <div class="modal-body">
+                <form action="{{route('ot.create')}}" method="POST">
+                    @csrf
+                    <div class="form-group">
+                        <label for="inputname">Select Date:</label>
+                        <input type="date" class="form-control" id="inputdate" name="inputdate" value="" required>
+                    </div>
+                    <div class="text-center">
+                        <button type="submit" class="btn btn-primary">CREATE</button>
+                    </div>
+                </form>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-default" data-dismiss="modal">CLOSE</button>
+            </div>
+        </div>
+    </div>
+</div>
+
 <div id="delOT" class="modal fade" role="dialog">
     <div class="modal-dialog">
         <div class="modal-content">
@@ -70,7 +95,7 @@
             <div class="modal-body text-center">
                 <div class="glyphicon glyphicon-warning-sign" style="color: #F0AD4E; font-size: 32px;"></div>
                 <p>Are you sure you want to delete claim for date <span id="deldate"></span>?<p>
-                <form action="{{ route('ot.remove') }}" method="POST">
+                <form action="{{ route('ot.delete') }}" method="POST">
                     @csrf
                     <input type="text" class="hidden" id="delid" name="delid" value="" required>
                     <button type="submit" class="btn btn-primary">DELETE</button>
@@ -90,6 +115,24 @@ $(document).ready(function() {
     $('#tOTList').DataTable({
         "responsive": "true",
     });
+});
+
+$('#newOT').on('show.bs.modal', function() {
+    var dt = new Date();
+    var m = dt.getMonth()+1;
+    if(m < 10){
+        m = "0"+m;
+    }
+    d = dt.getDate().toString();
+    while(d.length<2){
+        d = "0"+d;
+    }
+    $("#inputdate").val(dt.getFullYear()+"-"+m+"-"+d);
+    $("#inputdate").attr("max", dt.getFullYear()+"-"+m+"-"+d);
+    
+
+    // $("#inputdatestart").val(dt.getFullYear()+"-"+m+"-"+dt.getDate()+"T"+dt.getHours()+":"+dt.getMinutes());
+    // $("#inputdateend").val(dt.getFullYear()+"-"+m+"-"+dt.getDate()+"T"+dt.getHours()+":"+dt.getMinutes());
 });
 
 $('#delOT').on('show.bs.modal', function(e) {
